+++ src/__tests__/index.js
@@ -15,25 +15,29 @@ expect.addSnapshotSerializer({
   },
 })
 
+const shouldError = code => ({ code, error: true })
+
 pluginTester({
   plugin,
   snapshot: true,
-  babelOptions: { filename: __filename },
+  babelOptions: {
+    presets: ['react'],
+  },
   tests: [
     `
-      const MySource = ''
+      let MySource
       function ScopedFunction () {
         /* @source MySource */ const foo = 'bar' /* @source MySource */
       }
     `,
     `
       function ScopedFunction () {
-        const MySource = ''
+        let MySource
         /* @source MySource */ const foo = 'bar' /* @source MySource */
       }
     `,
     `
-      const MySource = ''
+      let MySource
       /* @source MySource */ const foo = 'bar' /* @source MySource */
     `,
     `
@@ -49,7 +53,7 @@ pluginTester({
       // @source MySource
     `,
     `
-      const MySource = ''
+      let MySource
       // @source MySource
       const MyComponent = () => (
         console.log('Hello!')
@@ -71,7 +75,7 @@ pluginTester({
       // @source MyOtherSource
     `,
     `
-      const MySource = ''
+      let MySource
       // @source MySource
       // @source MySource
     `,
@@ -88,9 +92,9 @@ pluginTester({
       let MySource = 'I should be the here'
       /* other comment */ const foo = 'bar' /* other comment */
     `,
-    `
+    shouldError(`
       /* @source MySource */ const foo = 'bar' /* @source MySource */
-    `,
+    `),
     `
       /* other comment */ const foo = 'bar' /* other comment */
     `,
@@ -131,12 +135,82 @@ pluginTester({
         // @source MySource
       )
     `,
+    shouldError(`
+      const MyComponent = () => (
+        // @source MySource
+        console.log('Hello!')
+        // @source MySource
+      )
+    `),
     `
-    const MyComponent = () => (
-      // @source MySource
-      console.log('Hello!')
-      // @source MySource
-    )
-  `,
+      const MySource = ''
+  
+      const MyComponent = () => (
+        // @source MySource
+        console.log('Hello!')
+        // @source MySource
+      )
+    `,
+    `
+      let MyJSXSource
+      const MyComponent = () => (
+        // @source MyJSXSource
+        <div>Hello!</div>
+        // @source MyJSXSource
+      )
+    `,
+    `
+      let sourceCode
+
+      console.log(sourceCode)
+
+      export default () => (
+        <Sidebar>
+          <ChartConfig dataType="time">
+            {({ data }) => (
+              // @source sourceCode
+              <Chart data={data}>
+                <Axis primary type="time" position="bottom" />
+                <Axis type="linear" position="left" stacked />
+                <Series type={Area} />
+                <Tooltip />
+              </Chart>
+              // @source sourceCode
+            )}
+          </ChartConfig>
+          <Code source={sourceCode} />
+        </Sidebar>
+      )
+    `,
+    `import React from 'react'
+
+      //
+      
+      import Sidebar from 'components/Sidebar'
+      import ChartConfig from 'components/ChartConfig'
+      import Code from 'components/Code'
+      
+      import { Chart, Axis, Series, Tooltip, Area } from '../../../src'
+      
+      const sourceCode = ''
+      
+      export default () => (
+        <Sidebar>
+          <ChartConfig dataType="time">
+            {({ data }) => (
+              // @source sourceCode
+              <Chart data={data}>
+                <Axis primary type="time" position="bottom" />
+                <Axis type="linear" position="left" stacked />
+                <Series type={Area} />
+                <Tooltip />
+              </Chart>
+              // @source sourceCode
+            )}
+          </ChartConfig>
+          <Code source={sourceCode} />
+        </Sidebar>
+      )
+    `,
   ],
 })