+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.0.2
+
+* Source strings will no longer contain generated code from babel (eg. semi-colons, etc).
+* Source strings now truncate lines to the shortest indentation
+
 ## 0.0.1
 
 Initial Commit. Yay!